+++ apps/web/src/app/(authenticated)/integrations/components/integrations-grid.tsx
@@ -3,7 +3,6 @@
 import { useSearchParams } from 'next/navigation';
 import Image from 'next/image';
 import { SimpleGrid, Transition } from '@mantine/core';
-import { notifications } from '@mantine/notifications';
 import { useCallback, useEffect, useState, type ReactNode } from 'react';
 import {
   type IntegrationsQuery,
@@ -72,15 +71,6 @@ export default function IntegrationsGrid(props: IntegrationProps): ReactNode {
     setTimeout(() => {
       setIsMounted(true); // Set isMounted to true after a delay to allow the transition to play
     }, 0);
-
-    // Show error notification
-    const error = searchParams.get('error');
-    if (error) {
-      notifications.show({
-        message: error,
-        color: 'red',
-      });
-    }
   }, [searchParams, updateIntegrationsData]);
 
   const isIntegrationAvailable = (status: IntegrationStatus): boolean => status !== IntegrationStatus.ComingSoon;
